Added basic implementation for cmd args parsing
The program now has a basic structured used to implement a basic command line argument parsing system.

A new class has been created to store the data of all of the existing arguments, each with a function pointer to the function that they use under the hood.

WIP : Implement file packing and unpacking, as well as path packing and unpacking.