Updated arg parsing
Arg parsing now notifies the user when unexpected arguments were issued, as well as when not enough arguments were found.

The help command now does not execute multiple times if the --help or -h args are passed multiple times.

Also, the usage of the help command prevents any code from being executed, so all other args are parsed for correctness but ignored afterward. The program prints the help and then quits.