Separated command registering from execution
Command line argument parsing is now separated in 2 steps for correctness and ease of use, as well as making the behaviour more consistent and predictable from the user side.

The args are processed in 2 steps:

Step 1: argument parsing. Checks the input arguments, registers them for execution and makes sure that they are correct before running them.

Step 2: command execution. The registered commands are executed, meaning that the registered commands are safe to run. Any errors that appear from this point onward are execptions thrown by the packer and unpacker parts of the programs, which would come from errors or wrong information located within the input files.