Replaced Func<>s with Action<>s
The code no longer returns the execution status from the Cmd() functions, so the code has returned to using Actions rather than Funcs for the function pointers.

Cmd methods are now responsible only of registering commands and thus have no failing status.

Exec methods are now the ones in charge of actually handling code execution and possible failures and exceptions.

All error handling for the argument parsing is done within the TryParseCommands and TryRunCommand, which should probably be renamed to TryPraseArgs and TryRegisterArg or something like that.