Added directory based packing and unpacking
The packing and unpacking commands now support directory based operations.

If the input of the -p and -u commands is a directory, the output is assumed to be a directory name.

The input directory is recursively searched through. All of the files within each directory is added to be packed and unpacked. All of the subdirectories are added to be created on the output directory path.

The directories are generated before the files to ensure that the output path directory structure is fully generated before performing any file creation.