Created MagickaLoadException class
The MagickaLoadException class is an Exception class that is used to handle errors when loading data that would not be valid within Magicka when loaded but that we cannot filter out through JSON alone.

An example of where this is useful is when loading data for CharacterTemplates, such as defining an input JSON with a list of lights with more than 4 point lights in total. That is not valid and actually generates a runtime exception within Magicka's code, so we can do the same to prevent us from compiling XNBs that will not work in the first place.

Note that this could be modified in the future and completely discarded if the engine rewrite project ends up allowing extensions that remove some of these limits... but that's a project for the far future and should not limit the correctness of this project since it does not even exist in the first place.

NOTE : Similar to how version specific data structures loading will be implemented with a flag for the packer/unpacker program, we can also define a flag for the engine rewrite support in the future, so this is alright for now.