Created Magicka Events classes
The Events.cs file contains all of the classes for each of the possible Magicka Events that exist within the game.

TODO : Implement each event type, making each hold their corresponding data and the information that they require.

TODO : In the future, we may repurpose the specific type of this class from within the JSON file to prevent having to type the $type string for JSON serializer as well as the enum value for MagickaPUP.