Created EventStorage class
The event storage class is responsible for holding the data of a specific event in memory.

It contains an event type enum and a Magicka Event property.

NOTE : Read the TODOs on the previous commit to remove duplication of identifying strings and enum values when working with .NET's json serializer. Maybe the solution could be to add a static field to each specific class that returns the enum value for said type identifier, and that way we no longer hold the value within the EventStorage class itself. But if we're going to do that, we might as well modify the code so that we don't hold the length of lists on a separate variable within the JSON...