Created Damage struct
The Damage struct contains the properties / fields required to describe a type of damage and the amount of damage to be applied to an entity.

TODO : Rewrite some of the old CharacterTemplate related code to use this Damage struct rather than doing its own thing.

TODO : Figure out what to do when rewriting the whole XnaObject based system, because as you can see, mpup's philosophy of "everything can be a primary object within an XNB file" is turning out to be quite a pain in the ass to maintain, as most certainly, even if that were true, certain objects just have inconsistent writing and reading orders, thus those would CLEARLY NEVER be used as primary objects within a file. Even if the objective is to offer as much extensibility as possible and support for future modded versions of the engine with custom behaviours and structures, maybe some limits should be applied.