Modified CharacterTemplate's enums to use Json String Enum converters
The JSON files will now use string serialized comma separated lists for enum values.

Enums that can take a single value will be treated as a single string.

Enums that are marked as flags will be capable of taking the form of multiple values through a single string with comma separated entries.