Created AbilityStorage class
The ability storage class is a proxy class that moves the primary generic reading logic out of the Ability class and allows making the reading and writing process easier to deal with.

NOTE : The Ability class must be modified and turned into an abstract class from which all of the specific ability types inherit from.