Implemented new read logic for AbilityStorage
The new ability storage system requires changing a bit the order of execution of the instructions to make sure that the bytes from the input XNB file are read in the correct order.

The new class hierarchy with the AbilityStorage class has been made so that the JSON serialization can be easily performed and to make the code easier to handle and maintain in the long run.