Created new GetRequiredContentReaders virtual method for XnaObject
This method will allow the packer to get the required content type readers for a given primary object that is going to be written into an XNB file.

This allows picking what content type readers to write rather than just writing all of the known content type readers, which is unnecessary and could lead to binary bloat in the final XNB files in the long run as the number of content type readers supported by mpup increases.