Added content type reader data to XnbFileObject
This is a WIP.

The logic for content type readers is going to be moved to the JSON file section. This is not really the best idea, as it allows input JSON files to define the required content type readers, and it could lead to terribly malformed XNB files, but it does come with the benefit that while debugging we can more easily see what content type readers are present... this does make support for external tools like MagickCow harder since now they also need to hardcode content type reader strings, so this idea should probably be discarded. Specially since it does not have any benefits... because you cannot define custom or unimplemented content type readers, that still requires an extension of the base mpup tool.