Implemented lights reading for PhysicsEntityTemplate
NOTE : Magicka does not support lights for physics entities, but it was planned to add support at some point. The XNB files contain a single field for the number of lights for a physics entity, but an exception is thrown within Magicka's code when the number of lights found is greater than 0, so all we're doing here is enforcing that by throwing an exception too.