Created context classes
TODO : Rework the internal system to use the read and write contexts rather than passing around all of the required parameters as individual objects.

In short, replace the Reader / Writer + Debug logger passing mechanism.

NOTE : This new system will allow passing a list of content readers, which can be used to write down the required content readers as we read / write the file. Altough this implementation could be modified in the future depending on the needs of the project.