Modified unpacker code to use memory stream and C# File operations rather than hand rolled streams
The memory stream will allow the binary reader to operate over a memory buffer rather than performing multiple file accesses during the read operations, which increases the performance of the reader at the cost of consuming more memory during runtime at a single time.

TODO : In the future, modify this code to read in chunks and perform some kind of paging if the file is too large to fit in memory?
Or maybe just don't read into a memory stream if possible. Altough, holding the entire data in memory is required when automatic decompression is implemented (unless we want to allow the program to write to a temporary file during LZX decompression when the feature is implemented, that should be a setting that one can change with a flag tho)