Commented out XNB header validation code
This code is going to be moved into the pup classes instead of being on the XnbFile class.

It could maybe make sense to have a separate class in charge of both the reading and writing process as well as the XNB header validation.

All of the XNB reading process begins on the XnbFile class, which means that whoever the top level class will be (as of now, the idea is for those to be the PUP classes), will be the one responsible for validating the XNB header and decompressing the XNB file, and then passing all of the data down to the XnbFile class so that the reading process can be entirely isolated from the decompression process, making it entirely transparent, and XNB validation gets to be performed only once, before performing decompression.