Modified debug logs in XnbFileData to match
The debug logs now notify of the fact taht we're reading the XNB Data stored within the XNB file itself rather than just reading the XNB file.

This is because XNB files need to be decompressed if they are compressed, so the XNB data reading is a step that takes place after decompression.