Removed NULL bytes padding logic from XnbFileData class
The NULL padding is performed as of now by the top level class XnbFile, so it no longer belongs in this class.

NOTE : Maybe in the future, if we add support for XnbFile compression when generating an output XNB file, it would probably make more sense for the padding bytes to either no longer exist or just be added at the end of the file data so that the compression would include them as well. That would mean we'd have to move the function back here and remove it from the top level class XnbFile