Implemented more complex arg parsing system with fallback values
This new system allows interpreting input values of the wrong type as a value that could be correct.

For instance, passing an integer as an input for a bool parse will cast the integer to bool.

The parsing tries to keep it similar to the way things work in C, but it is not 100% the same (for example, negative numbers are considered to be false for simplicity when dealing with floating point values) when in truth, any non 0 value should be considered true.