Created TypeReader class
This may not be the best idea, but this class is the base class for the ones that will be in charge of implementing the read and write methods for each type.

This will require a major rework of the logic where the reading and writing is moved from each specific object to readers.

The ReadInstance() and WriteInstance() methods will be completely needless and removed after this change is performed, if it is successful.

Also, this means that the ReadObject() and WriteObject() functions are most likely either going to be moved into an XNA Utility class or into the binary Writer and Reader classes.