Added new files to csproj and enabled unsafe code
Enabled unsafe code within the mpup project so that unsafe operations can be performed.

This is only required due to the D3D side of the code that is being worked on to allow shader decompilation and compilation to work without forcing users to require external tools, altough this pretty much kills full crossplatform support for the shader handling side of the code, since DirectX support is required.