Created FormHelper static class
The purpose of this class to aid with creating and destroying aux helper dummy forms.

The only reason this is being done is so that we can get a proper D3D9 device initialized, since D3D9 requires a valid window handle to be used for the device to be created and initialized.

Note that if a null pointer is passed, the current active window is apparently used as the window pointer for the device creation, which could lead to some behaviour that is not correct in any way, so this is why it's best to avoid said say of doing things.