Added fix for dictionary reader logic
Within XNA's code, the key reader and value reader are initialized at the start of the program.

We can't do that exactly due to how the whole program is structured (since I was not aware of this design choice until far later into development, so this needs to be refactored in the future), but we can still obtain the type reader instances on read every single time that Read is called.