Fixed issue with extension handling for input files on Packer
The GetExtension() function in C# from the Path library actually returns the extension with the dot (".") included, so the checks were wrong.

This is quite different from most other languages where the default path lib does not include the dot when retrieving the extension, which is where this mistake comes from. Whoops.