Updated Unpacker to remove unused variables, use new FSUtil function for output name generation and properly fixed leak by adding using() statement
Updated Unpacker to fix multiple known issues of the implementation.

The unused Context variable has been removed to clean up the implementation and decrease the unnecesarily bloated memory footprint of instancing a single Unpacker.

The new FSUtil function has been applied to generate output file names without unnecessarily adding noise to the WriteSystemFile() function.

A memory leak issue where the generated streams were not disposed of has been fixed by adding an using() statement around the read xnb file code. This used to be implemented with a .Dispose() call long ago, but at some point it had been accidentally removed. This has now been fixed.

TODO : Modify the main pup process (on the arg parser side of the code) so that we can have an UnpackerData / UnpackerConfig struct (same for Packer) and store those and pass them as argument to the Unpack() and Pack() methods so that the packer and unpacker can become static classes and we no longer need to needlessly bloat instances of this class with data that should only exist as a function argument...