Added JPG and BMP support
LOL Since bitmap already supports it generically, the code is as simple as adding the extension... the idea of this code as it is currently structure is quite literally to add support for sbtimage in the future, so that image formats not supported by C#'s System Bitmap class could be processed by stbimage instead... but for now, I guess I'll just add all of the image formats supported by bitmap and that's it.

Note that for now we could add very easily PPM support ourselves without much effort, so yeah.

TODO : Finish implementing the stb image support for true crossplatform support and wider image format support (remember that C#'s system drawing bitmap class only works on win32)