Temporarily disabled unused unsafe D3D and COM calls
This part of the code is going to be removed because a future implementation will be made with no win32 dependencies, which will allow crossplatform, non-win32 exclussive shader compilation and decompilation support.

So as to not remove the code to keep it around as a reference, and to allow release builds to exist without the /unsafe flag being used, the code has been temporarily disabled with a comment.