Modified pup cmd functions to only take path string args
Since the calls are sequential, the required args are already known at the time of the previous config point, and are stored within the PupProgram class instance itself. This allows the data to be passed around more trivially.

An alternative could be added for MT environments, if need be in the future, where a custom PupProgramSettings struct is passed around, containing config for all of the flags that the program can contain. Then, either convert that struct into the PackerSettings and UnpackerSettings structs (maybe add a getter for each within the PupProgramSettings struct itself for easier updating), or make it so that the Packer and Unpacker classes themselves take a global PupProgram config struct instead and then make it so that they use whichever params are useful to them.